Add HTML components to preview
