+++ src/components/App.js
@@ -1,11 +1,13 @@
 import React from 'react';
 // import '../scss/App.scss';
 import Collapsible from './Collapsible'
+import PreviewCard from './PreviewCard';
 
 function App() {
   return (
     <div className="App">
       <Collapsible/>
+      <PreviewCard/>
     </div>
   );
 }
